changed to portable-file-dialogs

+++ include/nimbus/core/application.hpp
@@ -81,7 +81,7 @@ class Application
         return m_drawLag;
     }
 
-    float getGameTime() const
+    double getGameTime() const
     {
         return m_gameTime;
     }
